+++ src/main/php/util/address/XmlStream.class.php
@@ -1,15 +1,16 @@
 <?php namespace util\address;
 
 use io\streams\InputStream;
-use lang\Value;
+use lang\{Closeable, Value};
 use util\Objects;
 
 /**
  * XML stream input
  *
+ * @test  util.address.unittest.XmlStreamTest
  * @test  util.address.unittest.XmlInputTest
  */
-class XmlStream extends Address implements Value {
+class XmlStream extends Address implements Closeable, Value {
   private $in;
 
   /**
@@ -26,12 +27,19 @@ class XmlStream extends Address implements Value {
 
   /** @return string */
   public function toString() {
-    return nameof($this).'<'.$this->in->toString().'>';
+
+    // Most InputStream instances have a toString() method but do not implement
+    // the Value interface, see https://github.com/xp-framework/core/issues/310
+    if ($this->in instanceof Value || method_exists($this->in, 'toString')) {
+      return nameof($this).'<'.$this->in->toString().'>';
+    } else {
+      return nameof($this).'<'.Objects::stringOf($this->in).'>';
+    }
   }
 
   /** @return string */
   public function hashCode() {
-    return 'S'.Objects::hashOf($this->file);
+    return 'S'.Objects::hashOf($this->in);
   }
 
   /**
@@ -42,5 +50,10 @@ class XmlStream extends Address implements Value {
    */
   public function compareTo($value) {
     return $value instanceof self ? Objects::compare($this->in, $value->in) : 1;
+  }
+
+  /** @return void */
+  public function close() {
+    $this->in->close();
   }
 }